Port the API functionality from uwsgi to FastAPI
*DO NOT MERGE THIS*

This is for review/demonstration purposes only. It was started as a clone of aladdin-demo for a
prototype project and some aspects can be merged back to aladdin-demo, but it can't be merged as is.

*Note:* This requires a new version of the `commands-base` container geared for `python3`.

This gives us all the bells and whistles of the `FastAPI` project, such as Swagger docs and an
`openapi.json` file. It is based on the `uvicorn` web application layer and we still keep nginx in
the mix for the webserver layer.

We have moved things over to use the new `asyncio` patterns where viable. The one notable exception
is the `elasticsearch` client. The async offerings there don't seem very maintained at the moment.

Other notable changes:
* We use the `python:3.8-slim` as the base docker image
* The project hieranrchy has been reorganized from the `app` directory to `components` with more
  clearly defined lines between the various aspects of the project like the `api` server and the
  `commands` app.
* There is a `shared` component that is leveraged by both the `api` and `commands` component. This
  is accomplished by using the `components` directory as the docker build context for all of the
  components. The components just need to pull in the pieces they need and ignore the rest. A later
  version of the `docker` client adds support for specifying a `.dockerignore` file at build time,
  so we can use that in the future to limit the context on a per-component basis.
* We mount this `components` directory when running with `--with-mount`
* The style git hook now runs in a docker container so we don't have to depend on or modify the
  local python installation
* The `build_docker.sh` script has been updated to adapt to this new hierarchy
* The python code has been cleaned up a bit
* We adopt the multi-stage Dockerfile pattern where possible
* We replace `print()` calls with `logger` calls