Get uvicorn --reload working
The auto-reloading only works when serving on a TCP port, not a domain socket. So we have to use
some configs to enable reloading only when started `--with-mount` and if the
`app.uvicorn.autoreload` value is `true`.

This also adds some upgrades to our Dockerfiles like pre-optimizing our python bytecode and
installing a `pip.conf` in the image.