Re-imagine the docker image build process
This is a pretty comprehensive refactoring of the `build/build_docker.sh` script. In fact, it's
replaced with a python script. The intent of this refactor is to provide a clean interface to
organize and define components and minimize the time needed to rebuild the docker images when code
or python library requirements are updated.

This new script leverages multi-stage docker builds and `FROM` injection to procedurally add layers
to an initial base image until you have all the desired contents. It also introduces the concept of
the `component.yaml` file. The new build script reads that file to determine how to build your
component image. While one may still just define a component with a simple `Dockerfile`, the new
build script allows you to specify an alternate base image and define which other components should
be included in your component's image, if any. In fact, all it takes to create a component image is
to create a directory under `components/`. Granted, it will only have `poetry` installed, but that's
enough to bootstrap a python component.

Component images built with the new process will be based on the current latest sanctioned python
image (`python:3.8-slim`) and have the poetry tool installed. Additionally, they will have the
`aladdin-user` and a workdir of `/code`. If the component or its dependency components define python
library requirements in their `pyproject.toml` file, those assets will be built and layered into the
image in modular fashion.

While the new script is focused on building python images, the hope is that with some moderate
effort, it could be extended to support projects in other languages, too.

These changes depend on an aladdin container update to provide two improvements:
 - Colorized output with more precise log levels
 - The ability to specify a subset of images to build with `aladdin build`. This is useful when
 iterating on a component's image and you don't want to waste time on other components to build,
 too.